Refactor : Login과 관련된 기능을 별도 클래스로 분리
